dlai/requirements.txt
# Modern Data Chatbot Requirements
# Core dependencies for hybrid tabular data chatbot

# Environment & Configuration
python-dotenv>=1.0.0             # Load environment variables from .env file

# Data Processing & Analytics
duckdb>=0.9.0                    # Ultra-fast OLAP database
pandas>=2.0.0                    # Data manipulation (compatibility)
numpy>=1.24.0                    # Numerical computing

# AI & ML
openai>=1.0.0                    # OpenAI API with structured outputs
sentence-transformers>=2.2.0     # Semantic embeddings for cache
pydantic>=2.5.0                  # Data validation and structured outputs

# Async & Performance
aiohttp>=3.8.0                   # Async HTTP requests
asyncio                          # Built-in async support

# Optional but recommended for production
# redis>=4.5.0                   # Production cache (replace in-memory cache)
# fastapi>=0.100.0               # Web API framework
# uvicorn>=0.20.0                # ASGI server
# prometheus-client>=0.16.0      # Metrics and monitoring